// server/src/proto/heimdallv2/checkpoint/tx.ts
export interface MsgCheckpoint {
  /** Address of the validator proposing this checkpoint. */
  proposer: string;
  /** First block number included in this checkpoint. */
  startBlock: number;
  /** Last block number included in this checkpoint. */
  endBlock: number;
  /** Merkle root hash of all blocks in the checkpoint range. */
  rootHash: Uint8Array;
  /** Merkle root hash of all account states in the checkpoint range. */
  accountRootHash: Uint8Array;
  /** Chain ID of the Bor chain. */
  borChainId: string;
}
// ...
export const MsgCheckpoint: MessageFns<MsgCheckpoint> = {
  encode(message: MsgCheckpoint, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.proposer !== "") {
      writer.uint32(10).string(message.proposer);
    }
    if (message.startBlock !== 0) {
      writer.uint32(16).uint64(message.startBlock);
    }
    if (message.endBlock !== 0) {
      writer.uint32(24).uint64(message.endBlock);
    }
    if (message.rootHash.length !== 0) {
      writer.uint32(34).bytes(message.rootHash);
    }
    if (message.accountRootHash.length !== 0) {
      writer.uint32(42).bytes(message.accountRootHash);
    }
    if (message.borChainId !== "") {
      writer.uint32(50).string(message.borChainId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): MsgCheckpoint {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMsgCheckpoint();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.proposer = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.startBlock = longToNumber(reader.uint64());
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.endBlock = longToNumber(reader.uint64());
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.rootHash = reader.bytes();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.accountRootHash = reader.bytes();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.borChainId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): MsgCheckpoint {
    return {
      proposer: isSet(object.proposer) ? globalThis.String(object.proposer) : "",
      startBlock: isSet(object.startBlock)
        ? globalThis.Number(object.startBlock)
        : isSet(object.start_block)
        ? globalThis.Number(object.start_block)
        : 0,
      endBlock: isSet(object.endBlock)
        ? globalThis.Number(object.endBlock)
        : isSet(object.end_block)
        ? globalThis.Number(object.end_block)
        : 0,
      rootHash: isSet(object.rootHash)
        ? bytesFromBase64(object.rootHash)
        : isSet(object.root_hash)
        ? bytesFromBase64(object.root_hash)
        : new Uint8Array(0),
      accountRootHash: isSet(object.accountRootHash)
        ? bytesFromBase64(object.accountRootHash)
        : isSet(object.account_root_hash)
        ? bytesFromBase64(object.account_root_hash)
        : new Uint8Array(0),
      borChainId: isSet(object.borChainId)
        ? globalThis.String(object.borChainId)
        : isSet(object.bor_chain_id)
        ? globalThis.String(object.bor_chain_id)
        : "",
    };
  },

  toJSON(message: MsgCheckpoint): unknown {
    const obj: any = {};
    if (message.proposer !== "") {
      obj.proposer = message.proposer;
    }
    if (message.startBlock !== 0) {
      obj.startBlock = Math.round(message.startBlock);
    }
    if (message.endBlock !== 0) {
      obj.endBlock = Math.round(message.endBlock);
    }
    if (message.rootHash.length !== 0) {
      obj.rootHash = base64FromBytes(message.rootHash);
    }
    if (message.accountRootHash.length !== 0) {
      obj.accountRootHash = base64FromBytes(message.accountRootHash);
    }
    if (message.borChainId !== "") {
      obj.borChainId = message.borChainId;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<MsgCheckpoint>, I>>(base?: I): MsgCheckpoint {
    return MsgCheckpoint.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<MsgCheckpoint>, I>>(object: I): MsgCheckpoint {
    const message = createBaseMsgCheckpoint();
    message.proposer = object.proposer ?? "";
    message.startBlock = object.startBlock ?? 0;
    message.endBlock = object.endBlock ?? 0;
    message.rootHash = object.rootHash ?? new Uint8Array(0);
    message.accountRootHash = object.accountRootHash ?? new Uint8Array(0);
    message.borChainId = object.borChainId ?? "";
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}